requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.2
gunicorn==21.2.0
